TheLastCEO/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
redis==5.0.1
celery==5.3.4 
djangorestframework-simplejwt==5.3.1
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
Pillow==10.1.0 
dj_database_url
uvicorn